+++ core/db.php
@@ -34,6 +34,7 @@ try{
   $dns=((!empty($settings['database']['driver']))?($settings['database']['driver']):'').((!empty($settings['database']['host']))?(':host='.$settings['database']['host']):'').((!empty($settings['database']['port']))?(';port='.$settings['database']['port']):'').((!empty($settings['database']['schema']))?(';dbname='.$settings['database']['schema']):'');
   $db=new PDO($dns,$settings['database']['username'],$settings['database']['password']);
   $db->exec("set names utf8");
+//  $db->setAttribute(PDO::ATTR_ERRMODE,PDO::ERRMODE_WARNING);
   $db->setAttribute(PDO::ATTR_ERRMODE,PDO::ERRMODE_EXCEPTION);
   if(isset($getcfg)&&$getcfg==true){
     $config=$db->query("SELECT * FROM `".$prefix."config` WHERE id=1")->fetch(PDO::FETCH_ASSOC);