+++ core/view/cart.php
@@ -18,13 +18,7 @@ if($page['notes']!=''){
 		'/<print page=[\"\']?notes[\"\']?>/',
 		'/<\/?pagenotes>/'
 	],[
-		(isset($_SESSION['rank'])&&$_SESSION['rank']>899?
-		'<form id="note-form" target="sp" enctype="multipart/form-data" method="post" action="core/update.php">'.
-			'<input type="hidden" name="id" value="'.$page['id'].'">'.
-			'<input type="hidden" name="t" value="menu">'.
-			'<input type="hidden" name="c" value="notes">'.
-			'<textarea class="editable" name="da">'.rawurldecode($page['notes']).'</textarea>'.
-		'</form>':rawurldecode($page['notes'])),
+		rawurldecode($page['notes']),
 		''
 	],$html);
 }else